% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgetFile.R
\docType{class}
\name{budget}
\alias{budget}
\title{Reference class budget}
\format{\code{\link{R6Class}} object.}
\usage{
budget
}
\value{
object of \code{\link{R6Class}} with budget storage and interface
  methods
}
\description{
This class holds and manages accounts, transactions and categories. It can be
ititialized with special RDS file or by add* methods. It also allows to save
the state in RDS file. This file is a simple copy of private members.
}
\section{Methods}{

\describe{
  \item{\code{new(path = NULL)}}{Object creator. You can provide path to RDS file or initialize empty object}
  \item{\code{save(path)}}{Save budget to RDS file}

  \item{\code{addCategory(category)}}{Add transaction category(ies) to budget}
  \item{\code{deleteCategory(category)}}{Delete transaction category(ies) from budget}
  \item{\code{getCategories()}}{Return category vector}
  \item{\code{moveCategory(oldCategory, newCategory)}}{Move old category to new one}

  \item{\code{addAccount(account)}}{Add account(s) to budget}
  \item{\code{deleteAccount(account)}}{Delete account(s) from budget}
  \item{\code{getAccounts()}}{Return accounts vector}

  \item{\code{addTransaction(account, transaction)}}{Add transaction(s) to account}
  \item{\code{deleteTransaction(account, trIds)}}{Delete transaction(s) from account}
}
}

\examples{
myBudget <- budget$new()
\dontrun{
myBudget <- budget$new("path/to/budget.rds")}
}
\author{
Daniel Rodak
}
\keyword{datasets}
