% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileSelectInput.R
\name{choose.file}
\alias{choose.file}
\title{Choose a File Interactively}
\usage{
choose.file(default = NA, caption = NA)
}
\arguments{
\item{default}{which folder to show initially (ignored on Windows)}

\item{caption}{the caption on the selection dialog (ignored on Windows)}
}
\value{
A length one character vector, character NA if 'Cancel' was selected.
}
\description{
Display a file selection dialog
}
\details{
Uses an Apple Script/\link{choose.file} to display a file
  selection dialog. With \code{default = NA}, the initial file selection is
  determined by default behavior of the "choose file" system command.
  Otherwise, paths are expanded with \link{path.expand}.
}
