% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_budget.R
\name{create_budget}
\alias{create_budget}
\title{Creates a budget schedule}
\usage{
create_budget(schedule, start = Sys.Date(), end = start + 90, initial = 0)
}
\arguments{
\item{...}{A series of budget items separated by commas.}
}
\value{
The output of \code{create_item} is a budget item.
}
\description{
\code{create_item} returns a budget schedule.
}
\examples{
#####
# Common use cases
#

# I need to pay my bills!
paycheck <- create_item( name = "Paycheck"
                       , amount = 1050
                       , day = "1"
                       , recurring = TRUE
                       )
rent <- create_item( name = "Rent"
                   , amount = -800
                   , day = "1"
                   , recurring = TRUE
                   )
groceries <- create_item( name = "Groceries"
                        , amount = -200
                        , day = "5"
                        , recurring = TRUE
                        )
my_bills <- create_schedule(paycheck, rent, groceries)
my_budget <- create_budget(my_bills)
}

