% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_schedule.R
\name{create_schedule}
\alias{create_schedule}
\title{Creates a budget schedule}
\usage{
create_schedule(...)
}
\arguments{
\item{...}{A series of budget items separated by commas or a single list of budget items.}
}
\value{
The output of \code{create_schedule} is a budget item.
}
\description{
\code{create_schedule} returns a budget schedule.
}
\examples{
# Create a paycheck item
paycheck <- create_item( name = "Paycheck"
                       , amount = 1000
                       , day = "2016-01-01"
                       , recurring = "monthly"
                       )
# Create a rent item
rent <- create_item( name = "Rent"
                   , amount = -500
                   , day = "2016-01-05"
                   , recurring = "monthly"
                   )

# Create a schedule
my_schedule <- create_schedule(paycheck, rent)
# Inspect
my_schedule

# Using all items in the current environment
my_schedule <- create_schedule(get_items())
# Inspect
my_schedule
}

