% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eom}
\alias{eom}
\title{Function which gives end of month for date}
\usage{
eom(date)
}
\arguments{
\item{date}{date vector}
}
\description{
Function which gives end of month for date
}
\author{
Daniel Rodak
}
