% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_corporate_tax.R
\name{estimate_corporate_tax}
\alias{estimate_corporate_tax}
\title{Estimate corporate taxes}
\usage{
estimate_corporate_tax(
  corporate_tax_2019 = 4677000,
  corporate_tax_index = indicies$corp_tax
)
}
\arguments{
\item{corporate_tax_2019}{A numeric value. The corporate tax at year 2019.}

\item{corporate_tax_index}{A numeric vector. A vector from 2018 to 2030.}
}
\value{
Atabble. The municipality corporete taxes by municipality and year.
}
\description{
Estimate municipality corporate taxes
}
\examples{
corporate_tax <- estimate_corporate_tax(corp_tax_2019 = 4677000)
}
