% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_goverment_tranfer.R
\name{estimate_goverment_trasfer}
\alias{estimate_goverment_trasfer}
\title{Estimate central government transfers to municipality}
\usage{
estimate_goverment_trasfer(pop_data, gov_tranfers = indicies$gov_trans)
}
\arguments{
\item{pop_data}{A tibble}

\item{gov_tranfers}{A tibble transfer per person}
}
\value{
A tabble. central government transfers to municipality by year
}
\description{
Estimate central government transfers to municipality
}
\examples{
goverment_trans <- estimate_goverment_transer()
}
