% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_income_tax.R
\name{estimate_income_tax}
\alias{estimate_income_tax}
\title{Estimate municipality income taxes}
\usage{
estimate_income_tax(
  pop_data,
  tax_rate = 0.1975,
  earnings_index = indicies$earnings
)
}
\arguments{
\item{pop_data}{A tibble. Population like functions get_population_data and forecast-population return}

\item{tax_rate}{A numeric vector. Annual tax rate vector starting from 2021. The last value of vector will be repated.}

\item{inflation}{A numeric vector. Annual inflation vector starting from 2019 -> 2020. The last value of vector will be repated.}
}
\value{
Atabble. The municipality income taxes by municipality and year.
}
\description{
Estimate municipality income taxes
}
\examples{
income_tax <- estimate_income_tax()
}
