% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_property_tax.R
\name{estimate_property_tax}
\alias{estimate_property_tax}
\title{Estimate corporate taxes}
\usage{
estimate_property_tax(
  property_tax_2019 = 14596000,
  property_tax_index = indicies$prop_tax
)
}
\arguments{
\item{property_tax_index}{A numeric vector. A vector from 2018 to 2030.}

\item{propertry_tax_2019}{A numeric value. The corporate tax at year 2019.}
}
\value{
A tabble. The municipality corporete taxes by municipality and year.
}
\description{
Estimate municipality corporate taxes
}
\examples{
property_tax <- estimate_property_tax(corp_tax_2019 = 4677000)
}
