% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sh_costs.R
\name{estimate_sh_costs}
\alias{estimate_sh_costs}
\title{Estimate standardized social and health service ongoing costs per person}
\usage{
estimate_sh_costs(
  pop_data,
  mspi = indicies$mspi,
  needfactor = needfactor,
  standardized_sh_cost = standardized_sh_cost,
  adj = c()
)
}
\arguments{
\item{pop_data}{A tibble. Population like functions get_population_data and forecast-population return}

\item{needfactor}{A list. Internal data structure. See \code{\link{needfactor}}.}

\item{inflation}{A numeric vector. Annual inflation vector starting from 2019 -> 2020. The last value of vector will be repated.}
}
\value{
The mortality rates by municipality, year, age and sex.
}
\description{
Estimate standardized social and health service ongoing costs per person by municipality and year
}
\examples{
fert_data <- get_mortality_data()        # return mortality rates by municipalities, year, age and sex
fert_data <- get_mortality_data("KU186") # return mortality rates by year, age and sex in Jarvenpaa
}
