% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_populaton.R
\name{forecast_population}
\alias{forecast_population}
\title{Forecast population}
\usage{
forecast_population(
  muni_code = "KU186",
  netchangerate = 0.015,
  netchangecount = NA,
  year = 2019,
  adj = 1.061497
)
}
\arguments{
\item{muni_code}{A character vector}

\item{netchangerate}{A numeric vector. An absolute net change rate in population by years. If a forecasting period is longer than netchangerate, the last value of vector will be repeted.}

\item{netchangecount}{A numeric vector. An absolute net change in population by years. If a forecasting period is longer than netchange, the last value of vector will be repeted. If both netchangerate and netchangecount is dilivered, netchangecount is used.}

\item{year}{A numeric value. A base year. Forecasting starts at the of this year.}
}
\value{
A list on two tibble. The First is the end of the year population and the second is the average of the start and the end of year by municipality, year, age and sex.
}
\description{
Forecast population
}
\examples{
popul_data <- get_mortality_data(muni_code = "KU186",  netchangerate = c(0.015, 0.015, 0.02, 0.02, 0.015), year = 2019)) # return populations by year, age and sex in Jarvenpaa
}
