% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_list.R
\docType{data}
\name{function_list}
\alias{function_list}
\title{The list of the functions of municipalities}
\format{
A tibble
\describe{
  \item{function_list}{A tibble of functions}
}
}
\usage{
function_list
}
\description{
The list of the functions of municipalities
}
\keyword{datasets}
