% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cost_data.R
\name{get_cost_data}
\alias{get_cost_data}
\title{Get net ongoing costs}
\usage{
get_cost_data(year = 2018)
}
\arguments{
\item{year}{A numeric value}
}
\value{
A tibble
}
\description{
Get net ongoing costs by municipality and function
}
\examples{
cost_data <- get_cost_data() # return ongoing costs in 2018 by municipality and function
}
