% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_financial_statements.R
\name{get_financial_statements}
\alias{get_financial_statements}
\title{Nettokäyttökustannukset toiminnoittain}
\usage{
get_financial_statements(municipality = "*", year = "*", lazy_load = TRUE)
}
\arguments{
\item{year}{numero tai vektori. Käyvät arvot ovat 2015 - 2018}
}
\value{
A tibble
}
\description{
Nettokäyttökustannukset toiminnoittain
}
\examples{
cost_data <- get_cost_data() # Nettokäyttökustannukset

}
