% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mortality_data.R
\name{get_mortality_data}
\alias{get_mortality_data}
\title{Get mortality data from  the the Statistic Finnish population forecast}
\usage{
get_mortality_data(muni_code = "*")
}
\arguments{
\item{muni_code}{A character vector}
}
\value{
The mortality rates by municipality, year, age and sex.
}
\description{
Get mortality data from  the the Statistic Finnish population forecast
}
\examples{
fert_data <- get_mortality_data()        # return mortality rates by municipalities, year, age and sex
fert_data <- get_mortality_data("KU186") # return mortality rates by year, age and sex in Jarvenpaa
}
