% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population_data.R
\name{get_population_data}
\alias{get_population_data}
\title{Get population data from  the the Statistic Finnish}
\usage{
get_population_data(muni_code = "*", year = 2019)
}
\arguments{
\item{muni_code}{A character vector}

\item{year}{A character vector. A year or years.}
}
\value{
A list on two tibble. The First is the end of the year population and the second is the average of the start and the end of year by municipality, year, age and sex.
}
\description{
Get population data from  the the Statistic Finnish
}
\examples{
popul_data <- get_population_data()                             # return population in 2019 by municipalities, age and sex
popul_data <- get_mortality_data("KU186", year = c(2018, 2019)) # return populations by year, age and sex in Jarvenpaa
}
