% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_population_data.R
\name{group_population_data}
\alias{group_population_data}
\title{Population by age groups for social and heath service need factor calculation}
\usage{
group_population_data(pop_data)
}
\arguments{
\item{pop_data}{A tibble, population by municipality, year, age and sex (like get_population_data retunrs).}
}
\description{
Population by age groups for social and heath service need factor calculation
}
\examples{
grouped_popul_data <- group_population_data(popul_data)
}
