% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ifNull}
\alias{ifNull}
\title{Replace NULLs with \code{rep}}
\usage{
ifNull(x, rep)
}
\arguments{
\item{x}{object}

\item{rep}{replacement}
}
\value{
object with replaced NULLs
}
\description{
Replace NULLs with \code{rep}
}
\author{
Daniel Rodak
}
