% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.schedule.R
\name{is.schedule}
\alias{is.schedule}
\title{Checks if an object is a budget schedule.}
\usage{
is.schedule(x)
}
\arguments{
\item{x}{An object.}
}
\description{
\code{is.schedule} Checks if an object is a budget schedule.
}

