% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{namedVecToList}
\alias{namedVecToList}
\title{Transform named vector to list}
\usage{
namedVecToList(x)
}
\arguments{
\item{x}{named vector}
}
\value{
list with names equal to vector names
}
\description{
Transform named vector to list
}
\author{
Daniel Rodak
}
