% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/needfactor.R
\docType{data}
\name{needfactor}
\alias{needfactor}
\title{The need factors ja model of social and health function by minicipality}
\format{
A list:
\describe{
  \item{factors}{A tibble:}
  \itemize{
    \item{muni}{numeric code of municipality of Finland}
    \item{needfactor_2017}{the need factors of social and health services for 2017}
    \item{needfactor_2018}{the need factors of social and health services for 2018}
  }
  \item{factors_source}{A tibble: source of data}
  \item{coefficients}{A tibble: coefficients of fitted model}
  \item{residuals}{A tbble: The parts of the need factors that model cannot explain.}
  \itemize{
    \item{muni}{numeric code of municipality of Finland}
    \item{residual}{the parts of the need factors that model cannot explain}
  }
}
}
\source{
\url{https://thl.fi/fi/web/sote-uudistus/talous-ja-politiikka/kustannukset-ja-vaikuttavuus/rahoitus/tarvevakioidut-menot}
}
\usage{
needfactor
}
\description{
A dataset contains the need factors of social and health services by municipality for 2017 and 2018.
}
\keyword{datasets}
