% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readQIF.R
\name{parseQIFLine}
\alias{parseQIFLine}
\title{Parse single QIF line}
\usage{
parseQIFLine(x, aggSplits = FALSE)
}
\arguments{
\item{x}{character vector with QIF line between '^' delimiters}
}
\description{
Parse single QIF line
}
