% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.budget.R
\name{plot.budget}
\alias{plot.budget}
\title{Plots a budget}
\usage{
\method{plot}{budget}(x)
}
\arguments{
\item{x}{A budget.}
}
\value{
The output of \code{plot.budget} is a plot of a budget.
}
\description{
\code{plot.budget} plots a budget object.
}
\examples{
#####
# Common use cases
#

# I need to pay my bills!
paycheck <- create_item( name = "Paycheck"
                       , amount = 1050
                       , day = 1
                       , recurring = TRUE
                       )
rent <- create_item( name = "Rent"
                   , amount = -800
                   , day = 1
                   , recurring = TRUE
                   )
internet <- create_item( name = "Internet"
                       , amount = -100
                       , day = 15
                       , recurring = TRUE
                       )
my_bills <- create_schedule(paycheck, rent, internet)
my_budget <- create_budget(my_bills, initial=500)
plot(my_budget)
}

