% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.budget.R
\name{print.budget}
\alias{print.budget}
\title{Prints a clean version of a budget}
\usage{
\method{print}{budget}(x, ...)
}
\arguments{
\item{x}{A budget.}

\item{...}{Does nothing.  Included only to use the print generic.}
}
\description{
\code{print.budget} prints a clean version of a budget.
}

