% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.item.R
\name{print.item}
\alias{print.item}
\title{Prints a clean version of a budget item}
\usage{
\method{print}{item}(x)
}
\arguments{
\item{x}{A budget item.}

\item{...}{Does nothing.  Included only to use the print generic.}
}
\description{
\code{print.item} prints a clean version of a budget item.
}

