% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.schedule.R
\name{print.schedule}
\alias{print.schedule}
\title{Prints a clean version of a budget schedule}
\usage{
\method{print}{schedule}(x)
}
\arguments{
\item{x}{A budget schedule.}
}
\description{
\code{print.schedule} prints a clean version of a budget schedule.
}

