% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBank.R
\name{readBank}
\alias{readBank}
\title{Functions to load data from banks}
\usage{
readBank(file, bank, ...)
}
\arguments{
\item{file}{path to file with transaction data}

\item{bank}{character with bank name, see Details for supported banks}

\item{...}{optional arguments passed to \code{\link[utils]{read.table}}}
}
\description{
Functions to load data from banks
}
\details{
This functions are adjusted \code{\link[utils]{read.table}} function
to load specific CSV files generated by various banks. Banks currently supported:
\itemize{
\item{mBank}{\code{mbank}}
\item{Idea Bank}{\code{idea}}
}
}
\author{
Daniel Rodak
}
