% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileSelectInput.R
\name{readFilenameSelectInput}
\alias{readFilenameSelectInput}
\title{Read the value of a filenameSelectInput}
\usage{
readFilenameSelectInput(session, inputId)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}

\item{inputId}{The id of the input object}
}
\description{
Read the value of a filenameSelectInput
}
\details{
Reads the value of the text field associated with a
\code{filenameSelectInput} object that stores the user selected file path.
}
