% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileSelectInput.R
\name{readFileSelectInput}
\alias{readFileSelectInput}
\title{Read the value of a fileSelectInput}
\usage{
readFileSelectInput(session, inputId)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}

\item{inputId}{The id of the input object}
}
\description{
Read the value of a fileSelectInput
}
\details{
Reads the value of the text field associated with a
\code{fileSelectInput} object that stores the user selected directory path.
}
