% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readQIF.R
\name{readQIF}
\alias{readQIF}
\title{Load data from QIF file generated by MicroSoft Money}
\usage{
readQIF(file, aggSplits = FALSE)
}
\arguments{
\item{file}{path to file with QIF transaction data}

\item{aggSplits}{logical. Convinience parameter to handle special case from
Microsoft Money. When transaction is splitted to some categories which are
later moved to the same category transactions are not aggregated}
}
\description{
Currently only Bank and Cash type accounts are supported
}
\author{
Daniel Rodak
}
