% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\docType{data}
\name{report}
\alias{report}
\title{Reference class for storing report settings and plotting it}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
report
}
\description{
Reference class for storing report settings and plotting it
}
\examples{
\dontrun{
cats <- bdg$getCategories()
cats <- cats[grepl("Wydatki:", names(cats))]
expenses <- report$new(bdg, "Wydatki", "table", "ParCat", "Month",
                       bdg$getAccounts(), cats,
                       as.Date(c("2017-10-31", "2018-03-31")))}
}
\keyword{datasets}
