% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModal.R
\name{reportSettings}
\alias{reportSettings}
\title{Modal window for editing or adding report}
\usage{
reportSettings(newReport = TRUE, bdg, metaFiller = list())
}
\arguments{
\item{newReport}{logical indicating whether this is new report}

\item{bdg}{objerct of class \code{budget}}

\item{metaFiller}{list given by report$metaFiller() to fill fields in the form}
}
\description{
Modal window for editing or adding report
}
\author{
Daniel Rodak
}
