% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spl_to_r}
\alias{spl_to_r}
\title{Function which transform handsontable split table into data.frame}
\usage{
spl_to_r(hot)
}
\arguments{
\item{hot}{hot table to be transformed}
}
\description{
This function is customized version of \code{\link[rhandsontable]{hot_to_r}}
suitable only for split data in \link{budget} R6 class
}
\author{
Daniel Rodak
}
