% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitTrans.R
\name{splitTransactionUI}
\alias{splitTransactionUI}
\alias{splitTransaction}
\title{UI for splitting transaction module}
\usage{
splitTransactionUI(id, accBut)

splitTransaction(input, output, session, mainTable, mainTable_select,
  budgetCats, trigger)
}
\arguments{
\item{id}{module ID}

\item{accBut}{action button for split acceptation}

\item{input}{input}

\item{output}{output}

\item{session}{sesion}

\item{mainTable}{reactive; hot table with transactions}

\item{mainTable_select}{reactive; select object associated with hot table}

\item{budgetCats}{reactive; categories for current budget}

\item{trigger}{reactive; counter to observe idicating whether \mainTable was
added successfully}
}
\description{
UI for splitting transaction module

Server for splitting transaction module
}
\author{
Daniel Rodak

Daniel Rodak
}
