% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{switchNames}
\alias{switchNames}
\title{Function which switches names and vector elements}
\usage{
switchNames(x)
}
\arguments{
\item{x}{vector}
}
\description{
Function which switches names and vector elements
}
\author{
Daniel Rodak
}
