% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_item.R
\name{update_item}
\alias{update_item}
\title{Updates a budget item}
\usage{
update_item(item, name, amount, day, recurring)
}
\arguments{
\item{item}{The budget item to be updated.}

\item{name}{The new name for the budget item (if applicable).}

\item{amount}{The new amount for the budget item (if applicable).}

\item{day}{The new day for the budget item (if applicable).}

\item{recurring}{The new recurring status for the budget item (if applicable).}
}
\value{
The output of \code{update_item} is a new budget item with the specified updates.
}
\description{
\code{update_item} returns an updated budget item.
}
\examples{
# Create a paycheck item
paycheck <- create_item( name = "Paycheck"
                       , amount = 1000
                       , day = "2016-01-01"
                       , recurring = "monthly"
                       )
# Inspect
paycheck

# Update the paycheck amount
paycheck <- update_item( paycheck
                       , amount = 1500
                       )
# Inspect
paycheck
}

