% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileSelectInput.R
\name{updateFilenameSelectInput}
\alias{updateFilenameSelectInput}
\title{Change the value of a filenameSelectInput on the client}
\usage{
updateFilenameSelectInput(session, inputId, value = NULL, ...)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}

\item{inputId}{The id of the input object.}

\item{value}{A directory path to set}

\item{...}{Additional arguments passed to \link{\code{choose.filename}}. Only
used if \code{value} is \code{NULL}.}
}
\description{
Change the value of a filenameSelectInput on the client
}
\details{
Sends a message to the client, telling it to change the value of the
input object. For \code{filenameSelectInput} objects, this changes the value
displayed in the text-field and triggers a client-side change event. A
directory selection dialog is not displayed.
}
