% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget.R
\docType{class}
\name{budget}
\alias{budget}
\title{Reference class budget}
\format{\code{\link{R6Class}} object.}
\usage{
budget
}
\value{
object of \code{\link{R6Class}} with budget storage and interface
  methods
}
\description{
This class holds and manages accounts, transactions and categories. It can be
ititialized with special RDS file or by add* methods. It also allows to save
the state in RDS file. This file is a simple copy of private members.
}
\section{Methods}{

\describe{
  \item{\code{new(path = NULL)}}{Object creator. You can provide path to RDS file or initialize empty object}
  \item{\code{save(path)}}{Save budget to RDS file}

  \item{\code{addCategory(category, budgetCat = rep(private$budgetCats[1], length(category)))}}{Add transaction category(ies) to budget along with budget category}
  \item{\code{deleteCategory(category)}}{Delete transaction category(ies) from budget}
  \item{\code{getCategories()}}{Return category vector}
  \item{\code{moveCategory(oldCategory, newCategory)}}{Move old category to new one}
  \item{\code{updateSystemCategories()}}{Set system categories to valid values}

  \item{\code{addBudgetCategory(budgetCat)}}{Add transaction budget category(ies) to budget}
  \item{\code{deleteBudgetCategory(budgetCat)}}{Delete transaction budget category(ies) from budget}
  \item{\code{getBudgetCategories()}}{Return budget category vector}

  \item{\code{addAccount(account, initialBalance = rep(0, length(account)))}}{Add account(s) to budget}
  \item{\code{deleteAccount(account)}}{Delete account(s) from budget}
  \item{\code{renameAccount(account, newName)}}{Rename account}
  \item{\code{setAccountInitialBalance(account, initialBalance = rep(0, length(account)))}}{Set new account(s) initial balance(s)}
  \item{\code{getAccounts()}}{Return accounts vector}
  \item{\code{getAccountInitialBalances()}}{Return account initial balances vector}
  \item{\code{getAccountBalances()}}{Return account balances vector}

  \item{\code{addTransaction(account, transaction)}}{Add transaction(s) to account}
  \item{\code{deleteTransaction(account, trIds)}}{Delete transaction(s) from account}
  \item{\code{getTransactionTable(account)}}{Get transaction table for account}
}
}

\examples{
myBudget <- budget$new()
\dontrun{
myBudget <- budget$new("path/to/budget.rds")}
}
\author{
Daniel Rodak
}
\keyword{datasets}
