% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_item.R
\name{create_item}
\alias{create_item}
\title{Creates a budget item}
\usage{
create_item(name, amount, day, recurring)
}
\arguments{
\item{name}{A name describing the budget item.}

\item{amount}{The amount associated with the budget item.}

\item{day}{The day associated with the budget item.  See Details for more information.}

\item{recurring}{Whether or not the budget item is recurring.}
}
\value{
The output of \code{create_item} is a budget item.
}
\description{
\code{create_item} returns a budget item.
}
\examples{
#####
# Common use cases
#

# I need to pay my rent!
rent <- create_item( name = "Rent"
                   , amount = -666
                   , day = "1"
                   , recurring = TRUE
                   )
rent
}

