% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_item.R
\name{create_item}
\alias{create_item}
\title{Creates a budget item}
\usage{
create_item(name, amount, day, recurring)
}
\arguments{
\item{name}{A name describing the budget item.}

\item{amount}{The amount associated with the budget item.  Positive values result in addition
and negative values result in subtraction.}

\item{day}{The day associated with the budget item.  See Details for more information.}

\item{recurring}{Whether or not the budget item is recurring.}
}
\value{
The output of \code{create_item} is a budget item.
}
\description{
\code{create_item} returns a budget item.
}
\details{
An item's day argument must be one of these values:

\itemize{
  \item \strong{numeric}: An integer between 1 and 31 (depending on the month!)
  \item \strong{character}: "last" (last day of the month)
}
}
\examples{
# Create a paycheck item
paycheck <- create_item( name = "Paycheck"
                       , amount = 1000
                       , day = 1
                       , recurring = TRUE
                       )
# Inspect
paycheck

# Create a rent item
rent <- create_item( name = "Rent"
                   , amount = -500
                   , day = 5
                   , recurring = TRUE
                   )
# Inspect
rent
}

