% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_schedule.R
\name{create_schedule}
\alias{create_schedule}
\title{Creates a budget schedule}
\usage{
create_schedule(...)
}
\arguments{
\item{...}{A series of budget items separated by commas or a single list of budget items.}
}
\value{
The output of \code{create_schedule} is a budget item.
}
\description{
\code{create_schedule} returns a budget schedule.
}
\examples{
# Create a paycheck item
paycheck <- create_item( name = "Paycheck"
                       , amount = 1000
                       , day = 1
                       , recurring = TRUE
                       )
# Create a rent item
rent <- create_item( name = "Rent"
                   , amount = -500
                   , day = 5
                   , recurring = TRUE
                   )

# Create a schedule
my_schedule <- create_schedule(paycheck, rent)
# Inspect
my_schedule

# Using a list of items
my_schedule <- create_schedule(list(paycheck, rent))
# Inspect
my_schedule
}

