% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_schedule.R
\name{create_schedule}
\alias{create_schedule}
\title{Creates a budget schedule}
\usage{
create_schedule(...)
}
\arguments{
\item{...}{A series of budget items separated by commas.}
}
\value{
The output of \code{create_item} is a budget item.
}
\description{
\code{create_item} returns a budget schedule.
}
\examples{
#####
# Common use cases
#

# I need to pay my bills!
rent <- create_item( name = "Rent"
                   , amount = -800
                   , day = 1
                   , recurring = TRUE
                   )
internet <- create_item( name = "Internet"
                       , amount = -100
                       , day = 15
                       , recurring = TRUE
                       )
my_bills <- create_schedule(rent, internet)
my_bills
}

