% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fertility_data.R
\name{get_fertility_data}
\alias{get_fertility_data}
\title{Get fertility data from  the the Statistic Finnish population forecast}
\usage{
get_fertility_data(muni_code = "*")
}
\arguments{
\item{muni_code}{A character vector}
}
\value{
The fertility rates by municipality and age.
}
\description{
Get fertility data from  the the Statistic Finnish population forecast
}
\examples{
fert_data <- get_fertility_data()        # return fertility rates by municipalities and mother ages
fert_data <- get_fertility_data("KU186") # return fertility rates by mother ages in Jarvenpaa
}
