% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_items.R
\name{get_items}
\alias{get_items}
\title{Gets items in the current environment}
\usage{
get_items(pattern = ".", env = parent.frame())
}
\arguments{
\item{pattern}{A regular expression describing the names of objects to check.  Defaults to
all objects}

\item{env}{The environment to look for items.  Defaults to the calling environment.}
}
\value{
The output of \code{create_item} is a list of budget items matching the specified pattern.
}
\description{
\code{get_items} returns a list of budget items matching the specified pattern.
}
\examples{
# Create items
rent <- create_item( name = "Rent"
                   , amount = -800
                   , day = 1
                   , recurring = TRUE
                   )
internet <- create_item( name = "Internet"
                       , amount = -100
                       , day = 15
                       , recurring = TRUE
                       )

# Get all items
get_items()

# Get only items matching "rent"
get_items("rent")
}

