% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_migration_data.R
\name{get_migration_data}
\alias{get_migration_data}
\title{Get migration data from  the the Statistic Finnish}
\usage{
get_migration_data(muni_code, year)
}
\arguments{
\item{muni_code}{A character vector like "KU186".}

\item{year}{A character vector. Rates have calculates as average between year - 3 and year.}
}
\value{
In and out migration rates by municipality, age and sex.
}
\description{
Get migration data from  the the Statistic Finnish
}
\examples{
migr_data <- get_migration_data()        # return in and out migrations rates by municipalities, age, sex
migr_data <- get_migration_data("KU186") # return in and out migrations rates by age and sex in Jarvenpaa
}
