% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.budget.R
\name{plot.budget}
\alias{plot.budget}
\title{Plots a budget}
\usage{
\method{plot}{budget}(x, ...)
}
\arguments{
\item{x}{A budget.}

\item{...}{Other arguments to pass to plot().}
}
\value{
The output of \code{plot.budget} is a plot of a budget.
}
\description{
\code{plot.budget} plots a budget object.
}
\examples{
# Create a paycheck item
paycheck <- create_item( name = "Paycheck"
                       , amount = 1000
                       , day = 1
                       , recurring = TRUE
                       )
# Create a rent item
rent <- create_item( name = "Rent"
                   , amount = -500
                   , day = 5
                   , recurring = TRUE
                   )

# Create a schedule
my_schedule <- create_schedule(paycheck, rent)

# Create a budget
my_budget <- create_budget(my_schedule, initial=1000)
# Plot
plot(my_budget)
}

