% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readQIF.R
\name{readQIF}
\alias{readQIF}
\title{Load data from QIF file generated by MicroSoft Money}
\usage{
readQIF(file)
}
\arguments{
\item{file}{path to file with QIF transaction data}
}
\description{
Currently only Bank and Cash type accounts are supported
}
\author{
Daniel Rodak
}
