% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tr_to_r}
\alias{tr_to_r}
\title{Function which transform handsontable object into data.frame}
\usage{
tr_to_r(hot)
}
\arguments{
\item{hot}{hot table to be transformed}
}
\value{
data.frame
}
\description{
This function is customized version of \code{\link[rhandsontable]{hot_to_r}}
suitable only for transaction data in \link{budget} R6 class
}
\author{
Daniel Rodak
}
