% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_budget.R
\name{update_budget}
\alias{update_budget}
\title{Updates a budget}
\usage{
update_budget(budget, start, end, initial)
}
\arguments{
\item{budget}{The budget to be updated.}

\item{start}{The new start date for the budget.}

\item{end}{The new end date for the budget.}

\item{initial}{The new initial amount for the budget.}
}
\value{
The output of \code{update_budget} is a new budget with the specified updates.
}
\description{
\code{update_budget} returns an updated budget.
}
\examples{
# Create a paycheck item
paycheck <- create_item( name = "Paycheck"
                       , amount = 1000
                       , day = 1
                       , recurring = TRUE
                       )
# Create a rent item
rent <- create_item( name = "Rent"
                   , amount = -500
                   , day = 5
                   , recurring = TRUE
                   )

# Create a schedule
my_schedule <- create_schedule(paycheck, rent)
#' # Create a budget
my_budget <- create_budget(my_schedule, initial=1000)
# Inspect
my_budget

# Update the initial amount
my_budget <- update_budget(my_budget, initial=0)
# Inspect
my_budget
}

