% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_schedule.R
\name{update_schedule}
\alias{update_schedule}
\title{Updates a budget schedule}
\usage{
update_schedule(schedule, add, remove)
}
\arguments{
\item{schedule}{The budget schedule to be updated.}

\item{add}{A list of budget items to add to the schedule.}

\item{remove}{A list of budget items to remove from the schedule.}
}
\value{
The output of \code{update_schedule} is a new budget schedule with the specified updates.
}
\description{
\code{updates_schedule} returns an updated budget schedule.
}
\examples{
# Create a paycheck item
paycheck <- create_item( name = "Paycheck"
                       , amount = 1000
                       , day = 1
                       , recurring = TRUE
                       )
# Create a rent item
rent <- create_item( name = "Rent"
                   , amount = -500
                   , day = 5
                   , recurring = TRUE
                   )

# Create a schedule
my_schedule <- create_schedule(paycheck, rent)
# Inspect
my_schedule

# Add an internet item to the schedule
internet <- create_item( name = "Internet"
                       , amount = -100
                       , day = 15
                       , recurring = TRUE
                       )
my_schedule <- update_schedule(my_schedule, add=internet)
# Inspect
my_schedule

# Remove the internet item from the schedule
my_schedule <- update_schedule(my_schedule, remove=internet)
# Inspect
my_schedule
}

